requirements.txt
Flask==3.1.0
keras==3.8.0
lime==0.2.0.1
matplotlib==3.7.5
numpy==1.23.5
Pillow==10.3.0
scikit_learn==1.3.2
shap==0.41.0
tensorflow==2.12.1
tensorflow_intel==2.11.1
tqdm==4.66.3
Werkzeug==3.1.3
